chat-backend/requirements.txt
fastapi
uvicorn[standard]
httpx
jinja2
psycopg2-binary